# cmd/laas/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AcademicQualification:
    enum:
    - None
    - Class-X
    - Class-XII
    - Diploma
    - Graduate
    - Post-Graduate
    type: string
    x-enum-varnames:
    - AcademicQualificationNone
    - AcademicQualificationClassX
    - AcademicQualificationClassXII
    - AcademicQualificationDiploma
    - AcademicQualificationGraduate
    - AcademicQualificationPostGraduate
...
  models.Category:
    enum:
    - General
    - SC
    - ST
    - OBC
    - Other
    type: string
    x-enum-varnames:
    - CategoryGeneral
    - CategorySC
    - CategoryST
    - CategoryOBC
    - CategoryOther
...
  models.DocumentsRequired:
    properties:
      description:
        type: string
      name:
        description: Unique constraint for the document name
        type: string
      type:
        type: string
    type: object
...
  models.Eligibility:
    properties:
      academic_qualification:
        $ref: '#/definitions/models.AcademicQualification'
      age_max:
        type: integer
      age_min:
        type: integer
      category:
        $ref: '#/definitions/models.Category'
      created_at:
        type: string
      documents_required:
        items:
          $ref: '#/definitions/models.EligibilityDocumentMap'
        type: array
      gender:
        $ref: '#/definitions/models.Gender'
      id:
        type: integer
      income_limit:
        type: number
      updated_at:
        type: string
    type: object
  models.EligibilityDocumentMap:
    properties:
      document:
        $ref: '#/definitions/models.DocumentsRequired'
      id:
        type: integer
      is_mandatory:
        type: boolean
    type: object
  models.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  models.Gender:
    enum:
    - Male
    - Female
    - Other
    type: string
    x-enum-varnames:
    - GenderMale
    - GenderFemale
    - GenderOther
...
  models.PaginationMeta:
    properties:
      limit:
        example: 10
        type: integer
      next:
        example: /api/v1/schemes?limit=10&page=11
        type: string
      page:
        example: 10
        type: integer
      previous:
        example: /api/v1/schemes?limit=10&page=9
        type: string
      resource_count:
        example: 200
        type: integer
      total_pages:
        example: 20
        type: integer
    type: object
  models.Scheme:
    properties:
      amount:
        type: number
      application_link:
        type: string
      created_at:
        type: string
      description:
        type: string
      eligibility:
        $ref: '#/definitions/models.Eligibility'
      eligibility_id:
        description: foreign key to Eligibility
        type: integer
      end_date:
        type: string
      id:
        type: integer
      name:
        type: string
      start_date:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  models.SchemeResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/models.PaginationMeta'
    type: object
...
  models.SuccessResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact:
    email: daschayan8837@gmail.com
    name: chayan das
    url: chayandass.com
  description: Enter your username and password for basic authentication
  license:
    name: GPL-2.0-only
    url: https://opensource.org/licenses/GPL-2.0
  title: Beneficiary Manager API
  version: "3.0"
paths:
...
  /schemes:
    get:
      consumes:
      - application/json
      description: Fetch available schemes with pagination and filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Scheme name
        example: Scholar Scheme
        in: query
        name: name
        type: string
      - description: Scheme status
        example: upcoming
        in: query
        name: status
        type: string
      - description: Minimum amount
        example: 1000
        in: query
        name: min_amount
        type: number
      - description: Maximum amount
        example: 5000
        in: query
        name: max_amount
        type: number
      - description: Start after date (RFC3339 format)
        example: "2023-01-01T00:00:00Z"
        in: query
        name: start_after
        type: string
      - description: End before date (RFC3339 format)
        example: "2023-12-31T23:59:59Z"
        in: query
        name: end_before
        type: string
      - description: Eligible gender
        in: query
        name: gender
        type: string
      - description: Academic qualification
        in: query
        name: academic_qualification
        type: string
      - description: Income limit
        in: query
        name: income_limit
        type: number
      - description: Category
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SchemeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get Schemes with pagination and filtering
      tags:
      - scheme
  /schemes/{id}:
    get:
      consumes:
      - application/json
      description: Get a specific scheme by its ID
      parameters:
      - description: Scheme ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Scheme'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get Scheme By ID
      tags:
      - scheme
  /schemes/{id}/applications:
    get:
      consumes:
      - application/json
      description: Get all the applications for a scheme by its ID
      parameters:
      - description: Scheme ID
        in: path
        name: id
        required: true
        type: string
      - description: Application status
        example: submitted
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all the applications for a scheme
      tags:
      - scheme
  /schemes/{id}/status:
    get:
      consumes:
      - application/json
      description: Get the current status of a scheme by its ID
      parameters:
      - description: Scheme ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get Scheme Status
      tags:
      - scheme
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"